requirements.txt
langchain
langchain-core
langchain-community
langgraph
langchain-groq
tavily-python
faiss-cpu
transformers
sentence-transformers
pypdf
pytesseract
Pillow
torch --index-url https://download.pytorch.org/whl/cpu
python-dotenv
flask
reportlab
